Rename package & fix unit test
