+++ build.sbt
@@ -1,7 +1,7 @@
 name := "sbt-cpd"
 organization := "com.github.sbt"
 
-version := "2.0.0-M2"
+version := "2.0.0-M3-SNAPSHOT"
 
 sbtPlugin := true
 crossSbtVersions := Seq("0.13.16", "1.0.2")
