+++ src/main/scala-sbt-0.13/de/johoop/cpd4sbt/Compat.scala
@@ -1,4 +1,4 @@
-package de.johoop.cpd4sbt
+package com.github.sbt.cpd
 
 object Compat {
   val Process = sbt.Process
