+++ src/main/scala/com/github/sbt/cpd/settings/CpdOutputType.scala
@@ -10,13 +10,11 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package de.johoop.cpd4sbt.settings
+package com.github.sbt.cpd.settings
 
-object CpdReportType {
-  sealed abstract class ReportType(val name: String)
+object CpdOutputType {
+  sealed trait OutputType
 
-  case object Simple extends ReportType("Simple")
-  case object VS extends ReportType("VS")
-  case object XML extends ReportType("XML")
-  case object CSV extends ReportType("CSV")
+  case object File extends OutputType
+  case object Console extends OutputType
 }
