+++ src/sbt-test/sbt-cpd/simple/src/main/scala/de/johoop/Main.scala
